requirements.txt
fastapi
uvicorn[standard]
email-validator
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]